misc: stm401: Fix conversion macros.
The STM32_TO_HOST macro was casting to the wrong data type causing
values to be truncated. Both conversion macros now use the correct size
data type when casting the result value.

Reviewed-on: http://gerrit.mot.com/740163
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ include/linux/stm401.h
@@ -544,8 +544,8 @@ struct stm_response {
 
 /* The following macros are intended to be called with the stm IRQ handlers */
 /* only and refer to local variables in those functions. */
-#define STM16_TO_HOST(x) ((short) be16_to_cpu(*((u16 *) (stm401_readbuff+(x)))))
-#define STM32_TO_HOST(x) ((short) be32_to_cpu(*((u32 *) (stm401_readbuff+(x)))))
+#define STM16_TO_HOST(x) ((int16_t) be16_to_cpu(*((u16 *) (stm401_readbuff+(x)))))
+#define STM32_TO_HOST(x) ((int32_t) be32_to_cpu(*((u32 *) (stm401_readbuff+(x)))))
 
 #define STM401_HALL_SOUTH 1
 #define STM401_HALL_NORTH 2
